refactored out the components
