+++ src/components/Browser.js
@@ -10,6 +10,7 @@ const BrowserFrame = styled(Box)`
   border-radius: 4px;
   border-top-width: 16px;
   box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
+  margin: -16px -3px -3px;
 `;
 
 const WindowControls = styled(Box)`
