+++ src/pages/projects/landhaus-altlengbach/_Screenshot.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Image from 'gatsby-image';
+
+const Screenshot = ({ image: { childImageSharp } = {}, ...rest }) => (
+  <Image {...childImageSharp} {...rest} />
+);
+
+export default Screenshot;
