+++ src/pages/projects/landhaus-altlengbach/index.js
@@ -1,78 +1,14 @@
 import React from 'react';
-import { Heading, Text, Flex, Box } from 'rebass';
+import { Box } from 'rebass';
 import { Link, graphql } from 'gatsby';
-import Image from 'gatsby-image';
-import styled, { css, keyframes } from 'styled-components';
-import { TransitionGroup, CSSTransition } from 'react-transition-group';
+import { css } from 'styled-components';
 
-import Layout from '../../components/layout';
-import Browser from '../../components/Browser';
-import Waypoint from 'react-waypoint';
-
-const Title = props => <Heading as="h1" fontSize={6} mb={3} {...props} />;
-const H2 = props => <Heading mb={3} {...props} />;
-const P = props => (
-  <Text
-    as="p"
-    mb={3}
-    css={`
-      max-width: 30em;
-    `} 
-    {...props}
-  />
-);
-
-const Col = props => <Box width={[1, 1 / 2]} px={[2, 3]} {...props} />;
-
-const IllustrationColumn = ({ children, ...props }) => (
-  <Col bg="#555" color="#fff" py={4} {...props}>
-    <Box css="height: 100%;">{children}</Box>
-  </Col>
-);
-const TextColumn = props => <Col bg="#fff" {...props} />;
-
-const Container = props => (
-  <Flex flexDirection={['column', 'row']} {...props} />
-);
-
-const WaypointAdapter = ({ children, innerRef }) => children(innerRef);
-
-const TextSection = ({ onSection, id, children, ...props }) => (
-  <Waypoint
-    bottomOffset="30%"
-    topOffset="30%"
-    onEnter={
-      onSection &&
-      (waypointProps =>
-        onSection(id, { type: 'enter', ...waypointProps, illustration: props }))
-    }
-    onLeave={
-      onSection &&
-      (waypointProps =>
-        onSection(id, { type: 'leave', ...waypointProps, illustration: props }))
-    }
-  >
-    <WaypointAdapter>
-      {ref => (
-        <Box
-          ref={ref}
-          py={4}
-          css={`
-            min-height: 100vh;
-          `}
-          id={id}
-          {...props}
-        >
-          {children}
-        </Box>
-      )}
-    </WaypointAdapter>
-  </Waypoint>
-);
-
-const Screenshot = ({ image: { childImageSharp } = {}, ...rest }) => (
-  <Image {...childImageSharp} {...rest} />
-);
+import Layout from '../../../components/layout';
+import { Title, H2, P, Container } from './_components';
+import { IllustrationColumn, TextColumn } from './_Column';
+import TextSection from './_TextSection';
+import Screenshot from './_Screenshot';
+import LayoutIllustration from './_LayoutIllustration';
 
 export default class AltlengbachPage extends React.Component {
   state = {
@@ -144,9 +80,7 @@ export default class AltlengbachPage extends React.Component {
                 currentSection={currentSection}
                 {...illustration}
                 href="https://altlengbach.netlify.com"
-              >
-                {illustration.renderContent && illustration.renderContent()}
-              </LayoutIllustration>
+              />
             </Box>
           </IllustrationColumn>
 
@@ -396,296 +330,3 @@ export const query = graphql`
     }
   }
 `;
-
-const LayoutSvg = styled.svg`
-  margin-left: auto;
-  max-width: 468px;
-  display: block;
-
-  && {
-    overflow: visible;
-  }
-`;
-
-const scaleX = keyframes`
-0%   { transform: scaleX(0); }
-5%   { transform: scaleX(0.5); }
-20%  { transform: scaleX(0.5); }
-25%  { transform: scaleX(1); }
-100% { transform: scaleX(1); }
-`;
-const scaleY = keyframes`
-0%   { transform: scaleY(0); }
-5%   { transform: scaleY(0.5); }
-15%  { transform: scaleY(0.5); }
-20%  { transform: scaleY(1); }
-100% { transform: scaleY(1); }
-`;
-
-const fadeIn = keyframes`
-0%  { opacity: 0; }
-20% { opacity: 1; }
-80% { opacity: 1; }
-85% { opacity: 0; }
-`;
-
-const Line = styled.path.attrs({ stroke: '#fff', strokeWidth: 2 })`
-  animation: ${({ speed = '10s', delay = '0s', horizontal, play }) =>
-    play
-      ? css`${horizontal ? scaleX : scaleY} ${speed} ${delay} ease infinite`
-      : 'none'};
-  transform: ${props =>
-    props.play ? (props.horizontal ? 'scaleX(0)' : 'scaleY(0)') : 'none'};
-  transform-origin: ${props => (props.horizontal ? '100% 0%' : '0% 0%')};
-`;
-
-const Rect = styled.rect.attrs({ fill: 'rgba(255, 255, 255, 0.3)' })`
-  animation: ${({ play, speed = '10s', delay = '0s' }) =>
-    play ? css`${fadeIn}  ${speed} ${delay} ease infinite` : 'none'};
-  opacity: ${props => (props.play ? 0 : 1)};
-  transition: ${props => (props.play ? 'none' : 'opacity 1s')};
-`;
-
-const moveBrowser = keyframes`
-0%      { transform: translateY(0%); }
-10%     { transform: translateY(100%); }
-26.667% { transform: translateY(100%); }
-36.667% { transform: translateY(0%); }
-53.333% { transform: translateY(0%); }
-63.333% { transform: translateX(-100%); }
-80%     { transform: translateX(-100%); }
-90%     { transform: translateX(0%); }
-`;
-
-const AnimatedGroup = styled.g`
-  animation: ${({ play, speed = '10s', delay = '0s' }) =>
-    play ? css`${moveBrowser}  ${speed} ${delay} ease infinite` : 'none'};
-
-  &.fade-enter,
-  &.fade-exit {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    transition: opacity 0.5s;
-  }
-  &.fade-enter {
-    opacity: 0;
-  }
-  &.fade-enter-active {
-    opacity: 1;
-  }
-  &.fade-exit-active {
-    opacity: 0;
-  }
-`;
-
-const Group = styled.g`
-  transform-origin: 100% 0;
-  transition: opacity 1s, transform 1s;
-  transform: ${({ center, zoom }) =>
-    `${center ? 'translate(-25%, 25%)' : ''} ${
-      zoom === 'canvas' ? 'scale(0.5)' : 'none'
-    }`};
-  opacity: ${({ show = true, opacity = 1 }) => (show ? opacity : 0)};
-`;
-
-const moveCanvas = keyframes`
-  0%      { transform: translateY(0%); }
-  10%     { transform: translateY(-100%); }
-  26.667% { transform: translateY(-100%); }
-  36.667% { transform: translateY(0%); }
-  53.333% { transform: translateY(0%); }
-  63.333% { transform: translateX(100%); }
-  80%     { transform: translateX(100%); }
-  90%     { transform: translateX(0%); }
-`;
-
-const Canvas = styled(Group)`
-  animation: ${({ play, speed = '10s', delay = '0s' }) =>
-    play ? css`${moveCanvas}  ${speed} ${delay} ease infinite` : 'none'};
-`;
-
-const LayoutIllustration = ({
-  zoom = '',
-  center = false,
-  width = 640,
-  height = 360,
-  showLines = false,
-  showCanvas = false,
-  showBrowser = false,
-  animate = '',
-  children,
-  classNames = 'fade',
-  contentKey,
-  timeout = 1000,
-  ...rest
-}) => (
-  <LayoutSvg viewBox={`0 0 ${width} ${height}`}>
-    <Group center={center} zoom={zoom} show>
-      <Canvas show={showCanvas} play={animate === 'canvas'} delay="2s">
-        <rect
-          x={-width}
-          y={0}
-          width={2 * width}
-          height={2 * height}
-          fill="rgba(0,0,0,0.1)"
-        />
-        <g className="contactSection">
-          <rect x={-width} y="0" width={width} height={height} fill="#6e8b3d" />
-          <foreignObject
-            x={-width}
-            y={height / 2 - height / 20}
-            width={width}
-            height={height / 10}
-            style={{ textAlign: 'center', fontSize: height / 10 }}
-          >
-            <p>Contact</p>
-          </foreignObject>
-        </g>
-        <g className="mainSection">
-          <rect x="0" y="0" width={width} height={height} fill="#52682d" />
-          <foreignObject
-            x={0}
-            y={height / 2 - height / 20}
-            width={width}
-            height={height / 10}
-            style={{ textAlign: 'center', fontSize: height / 10 }}
-          >
-            <p>Main</p>
-          </foreignObject>
-        </g>
-        <g className="plansSection">
-          <rect x={0} y={height} width={width} height={height} fill="#435626" />
-          <foreignObject
-            x={0}
-            y={height + height / 2 - height / 20}
-            width={width}
-            height={height / 10}
-            style={{ textAlign: 'center', fontSize: height / 10 }}
-          >
-            <p>Plans</p>
-          </foreignObject>
-        </g>
-
-        <Group show={showLines} opacity={showBrowser ? 0.2 : 1}>
-          <Line
-            play={animate === 'lines'}
-            delay="1s"
-            d={`M${(width / 13) * 5},0 L${(width / 13) * 5},${height * 2}`}
-          />
-          <Rect
-            delay="1.25s"
-            play={animate === 'lines'}
-            x="0"
-            y="0"
-            width={(width / 13) * 5}
-            height={height}
-          />
-          <Line
-            horizontal
-            delay="1.5s"
-            play={animate === 'lines'}
-            d={`M${width},${height * 0.85} L${-width},${height * 0.85}`}
-          />
-          <Rect
-            delay="1.75s"
-            play={animate === 'lines'}
-            x={0}
-            y={height * 0.85}
-            width={width}
-            height={height * 0.15}
-          />
-          <Line
-            horizontal
-            delay="2s"
-            play={animate === 'lines'}
-            d={`M${width},${height * 1.15} L${-width},${height * 1.15}`}
-          />
-          <Rect
-            delay="2.25s"
-            play={animate === 'lines'}
-            x={0}
-            y={height}
-            width={width}
-            height={height * 0.15}
-          />
-          <Rect
-            delay="2.75s"
-            play={animate === 'lines'}
-            x={0}
-            y={height}
-            width={(width / 13) * 5}
-            height={height}
-          />
-          <Line
-            delay="3s"
-            play={animate === 'lines'}
-            d={`M${width * -0.1},0 L${width * -0.1},${height * 2}`}
-          />
-          <Rect
-            delay="3.25s"
-            play={animate === 'lines'}
-            x={width * -0.1}
-            y="0"
-            width={width * 0.1}
-            height={height}
-          />
-          <Rect
-            delay="3.75s"
-            play={animate === 'lines'}
-            x={-width}
-            y={height * 0.85}
-            width={width}
-            height={height * 0.15}
-          />
-          <Rect
-            delay="4.25s"
-            play={animate === 'lines'}
-            x={-width}
-            y={height}
-            width={width}
-            height={height * 0.15}
-          />
-          <Rect
-            delay="4.75s"
-            play={animate === 'lines'}
-            x={width * -0.1}
-            y={height}
-            width={width * 0.1}
-            height={height}
-          />
-        </Group>
-      </Canvas>
-
-      <TransitionGroup component={null}>
-        {(children || showBrowser) && (
-          <CSSTransition
-            timeout={timeout}
-            classNames={classNames}
-            key={contentKey || 'default'}
-          >
-            <AnimatedGroup play={animate === 'browser'} delay="2s">
-              <foreignObject x={0} y={0} width={width} height={height}>
-                <Browser
-                  {...rest}
-                  css={css`
-                    ${children
-                      ? ''
-                      : `
-                        height: 100%;
-                        height: calc(100% + 19px);
-                      `};
-                    margin: -16px -3px -3px;
-                  `}
-                >
-                  {children}
-                </Browser>
-              </foreignObject>
-            </AnimatedGroup>
-          </CSSTransition>
-        )}
-      </TransitionGroup>
-    </Group>
-  </LayoutSvg>
-);
